requirements.txt
fastapi>=0.115.2,<1.0.0
mysql-connector-python
langchain-google-genai
langchain>=0.3.3,<0.4.0
langchain-core>=0.3.15,<0.4.0
langchain-community>=0.3.1,<0.4.0
pymongo
sqlalchemy
uvicorn
pydantic>=2.7.0,<3.0.0
protobuf